+++ openverse_catalog/dags/common/popularity/sql.py
@@ -4,7 +4,7 @@ from textwrap import dedent
 
 from common.constants import AUDIO, IMAGE
 from common.loader.sql import TABLE_NAMES
-from common.sql_helpers import PostgresHook
+from common.sql import PostgresHook
 from common.storage import columns as col
